Fix highlighting when page break syntax is used
When page breaks are used (e.g. "<<<") the remainder of the document
was turning yellow. It seems the page break syntax was triggering the syntax for internal links (e.g. "<<Jump Here>>").

This change removes the internal link styling aka 'Delimiter 7 style'.

Unless there's a smarter fix, I feel it's more important that the whole document *not* be yellow than having internal links that *are* yellow.